requirements.txt
paddlepaddle>=3.0.0
paddleocr>=2.7.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
easyocr>=1.7.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
google-generativeai>=0.3.0
streamlit>=1.28.0
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.66.0
matplotlib>=3.8.0
seaborn>=0.13.0
timm>=0.9.0
layoutparser>=0.3.4
detectron2>=0.6
